* Fix issue with fade out length: http://www.kdenlive.org:80/mantis/view.php?id=391 * Fix default direction of luma transition * Fix crash on transition deletion * Fix problem with transition resize
svn path=/branches/KDE4/; revision=2721

+++ src/gentime.cpp
@@ -41,7 +41,7 @@ double GenTime::ms() const {
 
 /** Returns the time in frames, after being given the number of frames per second */
 double GenTime::frames(double framesPerSecond) const {
-    return (int) floor(m_time * framesPerSecond + 0.5);
+    return floor(m_time * framesPerSecond + 0.5);
 }
 
 GenTime::~GenTime() {
